Removed some methods from cv::Algorithm and changed ml module interfaces
