+++ modules/face/src/precomp.hpp
@@ -43,12 +43,16 @@
 #ifndef __OPENCV_PRECOMP_H__
 #define __OPENCV_PRECOMP_H__
 
-#include "opencv2/face.hpp"
 #include "opencv2/imgproc.hpp"
 #include "opencv2/core.hpp"
 #include "opencv2/core/utility.hpp"
 #include "opencv2/core/private.hpp"
-
+#include "opencv2/core/persistence.hpp"
 #include <map>
+#include <iostream>
+#include <set>
+#include <limits>
+#include <iostream>
 
 #endif
+
