+++ modules/rgbd/samples/odometry_evaluation.cpp
@@ -173,13 +173,13 @@ int main(int argc, char** argv)
 
     Ptr<OdometryFrame> frame_prev = Ptr<OdometryFrame>(new OdometryFrame()),
                            frame_curr = Ptr<OdometryFrame>(new OdometryFrame());
-    Ptr<Odometry> odometry = Algorithm::create<Odometry>("RGBD." + string(argv[3]) + "Odometry");
+    Ptr<Odometry> odometry = Odometry::create("RGBD." + string(argv[3]) + "Odometry");
     if(odometry.empty())
     {
         cout << "Can not create Odometry algorithm. Check the passed odometry name." << endl;
         return -1;
     }
-    odometry->set("cameraMatrix", cameraMatrix);
+    odometry->setCameraMatrix(cameraMatrix);
 
     MyTickMeter gtm;
     int count = 0;
