+++ modules/xfeatures2d/include/opencv2/xfeatures2d/cuda.hpp
@@ -152,7 +152,7 @@ public:
     //! max keypoints = min(keypointsRatio * img.size().area(), 65535)
     float keypointsRatio;
 
-    GpuMat sum, mask1, maskSum, intBuffer;
+    GpuMat sum, mask1, maskSum;
 
     GpuMat det, trace;
