+++ modules/ximgproc/src/adaptive_manifold_filter_n.cpp
@@ -107,13 +107,18 @@ static void splitChannels(InputArrayOfArrays src, vector<Mat>& dst)
 class AdaptiveManifoldFilterN : public AdaptiveManifoldFilter
 {
 public:
-    AlgorithmInfo* info() const;
-
     AdaptiveManifoldFilterN();
 
     void filter(InputArray src, OutputArray dst, InputArray joint);
 
     void collectGarbage();
+
+    CV_IMPL_PROPERTY(double, SigmaS, sigma_s_)
+    CV_IMPL_PROPERTY(double, SigmaR, sigma_r_)
+    CV_IMPL_PROPERTY(int, TreeHeight, tree_height_)
+    CV_IMPL_PROPERTY(int, PCAIterations, num_pca_iterations_)
+    CV_IMPL_PROPERTY(bool, AdjustOutliers, adjust_outliers_)
+    CV_IMPL_PROPERTY(bool, UseRNG, useRNG)
 
 protected:
 
@@ -260,14 +265,6 @@ private: /*Parallelization routines*/
     };
 
 };
-
-CV_INIT_ALGORITHM(AdaptiveManifoldFilterN, "AdaptiveManifoldFilter",
-    obj.info()->addParam(obj, "sigma_s", obj.sigma_s_, false, 0, 0, "Filter spatial standard deviation");
-    obj.info()->addParam(obj, "sigma_r", obj.sigma_r_, false, 0, 0, "Filter range standard deviation");
-    obj.info()->addParam(obj, "tree_height", obj.tree_height_, false, 0, 0, "Height of the manifold tree (default = -1 : automatically computed)");
-    obj.info()->addParam(obj, "num_pca_iterations", obj.num_pca_iterations_, false, 0, 0, "Number of iterations to computed the eigenvector v1");
-    obj.info()->addParam(obj, "adjust_outliers", obj.adjust_outliers_, false, 0, 0, "Specify adjust outliers using Eq. 9 or not");
-    obj.info()->addParam(obj, "use_RNG", obj.useRNG, false, 0, 0, "Specify use random to compute eigenvector or not");)
 
 AdaptiveManifoldFilterN::AdaptiveManifoldFilterN()
 {
@@ -852,19 +849,17 @@ Ptr<AdaptiveManifoldFilter> AdaptiveManifoldFilter::create()
     return Ptr<AdaptiveManifoldFilter>(new AdaptiveManifoldFilterN());
 }
 
-CV_EXPORTS_W
 Ptr<AdaptiveManifoldFilter> createAMFilter(double sigma_s, double sigma_r, bool adjust_outliers)
 {
     Ptr<AdaptiveManifoldFilter> amf(new AdaptiveManifoldFilterN());
     
-    amf->set("sigma_s", sigma_s);
-    amf->set("sigma_r", sigma_r);
-    amf->set("adjust_outliers", adjust_outliers);
+    amf->setSigmaS(sigma_s);
+    amf->setSigmaR(sigma_r);
+    amf->setAdjustOutliers(adjust_outliers);
 
     return amf;
 }
 
-CV_EXPORTS_W
 void amFilter(InputArray joint, InputArray src, OutputArray dst, double sigma_s, double sigma_r, bool adjust_outliers)
 {
     Ptr<AdaptiveManifoldFilter> amf = createAMFilter(sigma_s, sigma_r, adjust_outliers);
